backend/pyproject.toml
[project]
name = "java-code-arena"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "flask>=3.1.2",
    "flask-cors>=6.0.2",
    "flask-socketio>=5.6.1",
    "javalang>=0.13.0",
    "nanoid>=2.0.0",
    "pillow>=12.1.1",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "waitress>=3.0.2",
]
